+++ app/Http/Controllers/TagsController.php
@@ -92,6 +92,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+       if($tag->posts->count() > 0 )
+       {
+           session()->flash('error','Tag can not be deleted becouse it has some posts.');
+
+           return  redirect()->back();
+       }
+
         $tag->delete();
 
         session()->flash('success','Tags Delete Successfully');
